Integrate standard pads(pads.c) into input.c
